MyPro1/requirements.txt
asgiref==3.4.1
author==0.0.0.0.dev20200926
cffi==1.15.1
cryptography==37.0.4
decorator==5.1.1
Django==3.2.15
django-crispy-forms==1.13.0
django-environ==0.9.0
django-filter==21.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
importlib-metadata==4.8.3
login==0.0.6
Markdown==3.3.7
myapp==0.1.dev0
pkg_resources==0.0.0
pycparser==2.21
PyJWT==v1.7.1
pytz==2022.2.1
self==2020.12.3
Serializer==0.2.1
simplejson==3.17.6
sqlparse==0.4.2
typing_extensions==4.1.1
zipp==3.6.0
